added table name collector overload for row id classes and added unit tests
